requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
anyio
playwright
python-dotenv

